Tidied up expressions API into core.
